fix: main page code refactor (hooks & structure)
