+++ frontend/app/lib/components/posts/PostsList.tsx
@@ -8,7 +8,6 @@ type PostsListProps = {
 export default function PostsList({ posts, userName }: PostsListProps) {
 	return (
 		<div className={styles.posts}>
-			<h2 className={styles.posts_hello}>Witaj, {userName}</h2>
 			{posts.map((post, index) => {
 				return <SinglePost post={post} key={index} />;
 			})}
