+++ frontend/app/page.module.scss
@@ -1,0 +1,8 @@
+.main {
+	max-width: var(--max-width);
+	margin: 0 auto;
+	padding-top: var(--nav-padding);
+	&_hello {
+		padding: 0 0 20px var(--global-padding);
+	}
+}
